requirements.txt
Flask
flask-pymongo
flask-marshmallow
pymongo
flask-bcrypt
python-dotenv
flask-restful
email-validator
